# OpenAPI.yml
openapi: 3.0.0
info:
  title: AgriAutomation API
  description: API for AgriAutomation
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /plantcaresystems:
    get:
      summary: Get a list of plant care systems
      tags:
        - Plant Care System
      operationId: getAllPlantCareSystems
      responses:
        '200':
          description: A list of plant care systems
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlantCareSystemResponse'
    post:
      summary: Create a new plant care system
      tags:
        - Plant Care System
      operationId: createPlantCareSystem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantCareSystemRequest'
            examples:
              WithoutMaintenanceTimeStamp:
                summary: Create PlantCareSystem without maintenanceTimeStamp
                value:
                  name: "Outdoor System"
                  description: "Outdoor plant monitoring system"
                  automationEnabled: false
      responses:
        '201':
          description: Plant care system created successfully
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantCareSystemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /plantcaresystems/{plantCareSystemId}:
    parameters:
      - $ref: '#/components/parameters/plantCareSystemId'
    get:
      summary: Get a plant care system by ID
      tags:
        - Plant Care System
      operationId: getPlantCareSystemById
      responses:
        '200':
          description: A single plant care system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantCareSystemResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a plant care system by ID
      tags:
        - Plant Care System
      operationId: updatePlantCareSystemById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantCareSystemRequest'
            examples:
              WithoutMaintenanceTimeStamp:
                summary: Update PlantCareSystem without maintenanceTimeStamp
                value:
                  name: "Updated System Name"
                  description: "Updated description"
                  automationEnabled: false
      responses:
        '200':
          description: Updated plant care system
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantCareSystemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      summary: Delete a plant care system by ID
      tags:
        - Plant Care System
      operationId: deletePlantCareSystemById
      responses:
        '204':
          description: Plant care system deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /plantcaresystems/{plantCareSystemId}/sensors:
    parameters:
      - $ref: '#/components/parameters/plantCareSystemId'
    get:
      summary: Get sensors for a plant care system
      tags:
        - Sensor
      operationId: getSensorsByPlantCareSystemId
      responses:
        '200':
          description: List of sensors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Add a sensor to a plant care system
      tags:
        - Sensor
      operationId: createSensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
            examples:
              WithoutTimestamps:
                summary: Create Sensor without reading and calibration timestamps
                value:
                  model: "SensorModelY"
                  temperature: 20
                  humidity: 50
      responses:
        '201':
          description: Sensor created successfully
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /plantcaresystems/{plantCareSystemId}/sensors/{sensorId}:
    parameters:
      - $ref: '#/components/parameters/plantCareSystemId'
      - $ref: '#/components/parameters/sensorId'
    get:
      summary: Get a specific sensor by ID
      tags:
        - Sensor
      operationId: getSensorById
      responses:
        '200':
          description: A single sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a sensor by ID
      tags:
        - Sensor
      operationId: updateSensorById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorRequest'
            examples:
              WithoutTimestamps:
                summary: Update Sensor without reading and calibration timestamps
                value:
                  model: "UpdatedModelZ"
                  temperature: 23
                  humidity: 57
      responses:
        '200':
          description: Updated sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      summary: Delete a sensor by ID
      tags:
        - Sensor
      operationId: deleteSensorById
      responses:
        '204':
          description: Sensor deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /plantcaresystems/{plantCareSystemId}/sensors/{sensorId}/plants:
    parameters:
      - $ref: '#/components/parameters/plantCareSystemId'
      - $ref: '#/components/parameters/sensorId'
    get:
      summary: Get plants for a specific sensor
      tags:
        - Plant
      operationId: getPlantsBySensorId
      responses:
        '200':
          description: List of plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlantResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Add a plant to a sensor
      tags:
        - Plant
      operationId: createPlant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantRequest'
            examples:
              CreatePlant:
                summary: Create a new plant
                value:
                  name: "Tomato"
                  growthStage: "Seedling"
      responses:
        '201':
          description: Plant created successfully
          headers:
            Location:
              description: URL of the created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /plantcaresystems/{plantCareSystemId}/sensors/{sensorId}/plants/{plantId}:
    parameters:
      - $ref: '#/components/parameters/plantCareSystemId'
      - $ref: '#/components/parameters/sensorId'
      - $ref: '#/components/parameters/plantId'
    get:
      summary: Get a specific plant by ID
      tags:
        - Plant
      operationId: getPlantById
      responses:
        '200':
          description: A single plant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a plant by ID
      tags:
        - Plant
      operationId: updatePlantById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantRequest'
            examples:
              UpdatePlant:
                summary: Update an existing plant
                value:
                  name: "Updated Tomato"
                  growthStage: "Vegetative"
      responses:
        '200':
          description: Updated plant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      summary: Delete a plant by ID
      tags:
        - Plant
      operationId: deletePlantById
      responses:
        '204':
          description: Plant deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    PlantCareSystemRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: Name of the plant care system
        description:
          type: string
          description: Description of the plant care system
        automationEnabled:
          type: boolean
          description: Automation status
      description: Request schema for creating or updating a PlantCareSystem

    PlantCareSystemResponse:
      allOf:
        - $ref: '#/components/schemas/PlantCareSystemRequest'
        - type: object
          properties:
            id:
              type: integer
              description: Unique identifier
              example: 1
          description: Response schema for PlantCareSystem including its ID

    SensorRequest:
      type: object
      required:
        - model
      properties:
        model:
          type: string
          description: Sensor model
        temperature:
          type: integer
          description: Current temperature reading
        humidity:
          type: integer
          description: Current humidity reading
      description: >
        Request schema for creating or updating a Sensor.
        The associated `plantCareSystemId` is provided via the URL path.

    SensorResponse:
      allOf:
        - $ref: '#/components/schemas/SensorRequest'
        - type: object
          properties:
            id:
              type: integer
              description: Unique identifier
              example: 1
            readingTimestamp:
              type: string
              format: date-time
              description: Timestamp of the sensor reading
              example: "2024-04-27T14:30:00"
            calibrationTimestamp:
              type: string
              format: date-time
              description: Last calibration timestamp
              example: "2024-04-01T12:00:00"
            plantCareSystemId:
              type: integer
              description: The unique identifier of the associated PlantCareSystem
              example: 1
          description: Response schema for Sensor including its ID and timestamps

    PlantRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the plant
        growthStage:
          type: string
          description: Growth stage of the plant (e.g., Seedling, Vegetative, Flowering)
      description: Request schema for creating or updating a Plant

    PlantResponse:
      allOf:
        - $ref: '#/components/schemas/PlantRequest'
        - type: object
          properties:
            id:
              type: integer
              description: Unique identifier
              example: 1
          description: Response schema for Plant including its ID

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: "Invalid request data"
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          example:
            - resource: "PlantCareSystem"
              field: "name"
              code: "must not be blank"
      description: Schema for error responses

    ErrorDetail:
      type: object
      properties:
        resource:
          type: string
          description: Resource where the error occurred
          example: "PlantCareSystem"
        field:
          type: string
          description: Field where the error occurred
          example: "name"
        code:
          type: string
          description: Error code
          example: "must not be blank"
      description: Detailed information about an error

  parameters:
    plantCareSystemId:
      name: plantCareSystemId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        example: 1
      description: The unique identifier of the plant care system
    sensorId:
      name: sensorId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        example: 1
      description: The unique identifier of the sensor
    plantId:
      name: plantId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        example: 1
      description: The unique identifier of the plant

  responses:
    BadRequest:
      description: Bad request due to invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Invalid request data"
            details:
              - resource: "PlantCareSystem"
                field: "name"
                code: "must not be blank"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "PlantCareSystem not found with id 1"
            details:
              - resource: "PlantCareSystem"
                field: "id"
                code: "not_found"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            message: "Validation Failed"
            details:
              - resource: "PlantCareSystem"
                field: "name"
                code: "size must be between 2 and 100"

tags:
  - name: Plant Care System
    description: Operations related to plant care systems
  - name: Sensor
    description: Operations related to sensors
  - name: Plant
    description: Operations related to plants